# Make (TF) CI faster (test only a random subset of model classes)