@HuggingFaceDocBuilderDev 2023-06-30
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/transformers/pr_24592). All of your documentation changes will be reflected on that endpoint.