@Rocketknight1 2023-06-30
Some of the very slow tests (like `test_saved_model_creation_extended`) only apply to a few models anyway - they're in `test_modeling_tf_core.py`, so they shouldn't have a big effect on the total test runtime. I might have a couple of ideas for speeding up `test_xla_fit` or `test_compile_tf_model`, though!