@ydshieh 2023-06-30
> Let's not take a random subset but the first two then. To test the base model and a model with head.

Would it be ok to take the first + a random other one with head?