@ydshieh 2023-06-30
@Rocketknight1 

But that test is not changed, i.e. it doesn't use `get_random_model_classes` introduced here.