@sgugger 2023-06-30
> Would it be ok to take the first one (base model) + a random other one with head?

I don't like randomness in tests as it makes them flaky.