@ydshieh 2023-06-30
Well, in this situation, I do prefer to keep a random head model.

  - We are reducing the number of model classes being tested due to the slow runtime. If we keep the fix model classes, we are likely to miss failures in certain model heads. (and for the involved tests in this PR, they all pass currently for their all model classes - if not, probably just one or two.)

  - Only slow tests are involved --> no flakyness shown on CircleCI.